Move BucketPQ definitions to header.
