+++ networkit/cpp/auxiliary/CMakeLists.txt
@@ -1,6 +1,5 @@
 networkit_add_module(auxiliary
     BloomFilter.cpp
-    BucketPQ.cpp
     Log.cpp
     NumericTools.cpp
     Parallelism.cpp
